requirements.txt
pywin32==306
customtkinter==5.2.2
pillow==10.4.0
requests==2.32.3